Move UndoLogControl struct into header to prepare for wider use.
A later commit will make use of it from other translation units so let's not
define it in undolog.c.  This requires hiding the definition when included
from FRONTEND code.  Perhaps this should have a new header of its own or some
other reorganization, but for now let's just do it conditionally.

Thomas Munro