Fix uur_next corruption by replacing global variables with UndoLogControl.
Previously we could corrupt our uur_next chain when the undo log you're
attached to changes.  This broke some later commits.

Get last_xact_start and prevlen in UndoLogControl instead of maintaining a
local copy.  They can be read directly from shmem without locking (but not
written) by the backend that is currently attached.

prev_txid remains as a global variable, and it needs to be cleared whenever
the undo log changes.  Perhaps it could be stored in UndoLogControl too, but
that leads to some circularities.

Thomas Munro, reviewed by XXX